requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.1
django-crispy-forms==1.14.0
django-filter==21.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.1.1
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
whitenoise==6.2.0
